requirements.txt
mcp>=1.2.0
prompt-toolkit>=3.0.47
PyYAML>=6.0.2
python-dotenv>=1.0.1
openai>=1.40.0
